+++ src/lib/plugins/MotionPlanReplayer/__tests__/parse-plan.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { parsePlan, PlanParseError } from './parse-plan'
+import { parsePlan, PlanParseError } from '../parse-plan'
 
 const MINIMAL_FRAME_SYSTEM = {
 	frames: {
